+++ Changelog.md
@@ -1,3 +1,15 @@
+### 0.8.2 (unreleased)
+
+Language Features:
+
+
+Compiler Features:
+
+
+Bugfixes:
+
+
+
 ### 0.8.1 (2021-01-27)
 
 Language Features: